requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-classic
langchain-openai
langchain-community
langchain-chroma
ipykernel
chromadb
docx2txt
PyPDF2
python-docx
ipykernel
PyPDF
psycopg2
python-jose
passlib
bcrypt==4.0.1
sqlalchemy
python-jose[cryptography]
dramatiq
redis
